自动选中小有问题

+++ src/components/TrashDetail.vue
@@ -54,6 +54,10 @@
       this.getTrashNotes()
         .then(() => {
           this.setCurTrashNote({ curTrashNoteId: this.$route.query.noteId })
+          this.$router.replace({
+            path: '/trash',
+            query: { noteId: this.curTrashNote.id }
+          })
         })
     },
 
@@ -83,15 +87,34 @@
       ]),
 
       onDelete() {
-        console.log({ noteId: this.curTrashNote.id })
-        this.deleteTrashNote({ noteId: this.curTrashNote.id })
+        this.$confirm('删除后将无法恢复', '确定删除吗？', {
+          confirmButtonText: '确定',
+          cancelButtonText: '取消',
+          type: 'warning'
+        }).then(() => {
+          return this.deleteTrashNote({ noteId: this.curTrashNote.id })
+        }).then(() => {
+          console.log('delete success')
+          this.setCurTrashNote()
+          this.$router.replace({
+            path: '/trash',
+            query: { noteId: this.curTrashNote.id }
+          })
+        })
       },
 
       onRevert() {
         this.revertTrashNote({ noteId: this.curTrashNote.id })
+          .then(() => {
+            this.setCurTrashNote()
+            this.$router.replace({
+              path: '/trash',
+              query: { noteId: this.curTrashNote.id }
+            })
+          })
       }
-
     },
+
 
     beforeRouteUpdate (to, from, next) {
       this.setCurTrashNote({ curTrashNoteId: to.query.noteId})
